requirements.txt
espn-api==0.27.0
pandas>=2.2
numpy>=1.26
requests>=2.32
python-dateutil>=2.9.0
pyyaml>=6.0.1
